# API/openapi/API_v2.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Mangette v2",
    "version": "2.0"
  },
  "paths": {
...
    "/v2/Settings/UserAgent": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get the current UserAgent used by Mangette",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Set a new UserAgent",
        "requestBody": {
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Settings"
        ],
        "summary": "Reset the UserAgent to default",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/v2/Settings/ImageCompressionLevel/{level}": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Set the Image-Compression-Level for Images",
        "parameters": [
          {
            "name": "level",
            "in": "path",
            "description": "100 to disable, 0-99 for JPEG ImageCompression-Level",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Level outside permitted range",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/Settings/BWImages/{enabled}": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Enable/Disable conversion of Images to Black and White",
        "parameters": [
          {
            "name": "enabled",
            "in": "path",
            "description": "true to enable",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v2/Settings/ChapterNamingScheme": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Gets the Chapter Naming Scheme",
        "description": "Placeholders:\r\n%M Obj Name\r\n%V Volume\r\n%C Chapter\r\n%T Title\r\n%A Author (first in list)\r\n%I Chapter Internal ID\r\n%i Obj Internal ID\r\n%Y Year (Obj)\r\n            \r\n?_(...) replace _ with a value from above:\r\nEverything inside the braces will only be added if the value of %_ is not null",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Sets the Chapter Naming Scheme",
        "description": "Placeholders:\r\n%M Obj Name\r\n%V Volume\r\n%C Chapter\r\n%T Title\r\n%A Author (first in list)\r\n%Y Year (Obj)\r\n            \r\n?_(...) replace _ with a value from above:\r\nEverything inside the braces will only be added if the value of %_ is not null",
        "requestBody": {
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v2/Settings/ConnectorPriority": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Connector names in download-failover order. The first source that has the chapter and is not cooling down is used.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Set connector download-failover order. Unknown names are ignored. Missing known connectors are appended.",
        "requestBody": {
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v2/Settings/FlareSolverr/Url": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Sets the FlareSolverr-URL",
        "requestBody": {
          "description": "URL of FlareSolverr-Instance",
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Settings"
        ],
        "summary": "Resets the FlareSolverr-URL (HttpClient does not use FlareSolverr anymore)",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v2/Settings/FlareSolverr/Test": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Test FlareSolverr",
        "responses": {
          "200": {
            "description": "FlareSolverr is working!"
          },
          "500": {
            "description": "FlareSolverr is not working"
          }
        }
      }
    },
...
    "/v2/Settings/DownloadLanguage/{Language}": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Sets the language in which Manga are downloaded",
        "parameters": [
          {
            "name": "Language",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v2/Settings/LibraryRefresh": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Sets the time when Libraries are refreshed",
        "requestBody": {
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/PatchLibraryRefreshRecord"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/PatchLibraryRefreshRecord"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/PatchLibraryRefreshRecord"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/PatchLibraryRefreshRecord"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LibraryRefreshSetting": {
        "enum": [
          "AfterAllFinished",
          "AfterMangaFinished",
          "AfterEveryChapter",
          "WhileDownloading"
        ],
        "type": "string"
      },
...
      "PatchLibraryRefreshRecord": {
        "required": [
          "setting"
        ],
        "type": "object",
        "properties": {
          "setting": {
            "$ref": "#/components/schemas/LibraryRefreshSetting"
          },
          "refreshLibraryWhileDownloadingEveryMinutes": {
            "type": "integer",
            "description": "When API.Workers.LibraryRefreshSetting.WhileDownloading is selected, update the time between refreshes",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}